requirements.txt
playwright==1.45.0
python-dotenv==1.0.1
supabase==2.6.0
pdfminer.six==20240706
tenacity==9.0.0
httpx==0.27.2
uvloop==0.19.0; sys_platform != 'win32'

